requirements.txt
pandas
numpy
google-cloud-storage
scikit-learn
pyyaml
matplotlib
seaborn